Revert "Factor out varint"
This reverts commit baa23738f8ec8253c484910cc8c8de95eb791509.
